Save ordering information for packets with repeated field keys
This change stores an index in the `LayerFieldsContainer` which indicates the order in which the fields existed in the raw packet. This is required for protocols such as OSC, which have field names that match the data type (and thus are repeated frequently) and which the order in which the fields existed in the packet is crucial to parsing the packet.